requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0

mccabe==0.7.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
sqlparse==0.5.2
typing_extensions==4.12.2
